Separated Utils fils into a separate script
